refactor: refactor UserMenu - move some of the elements to separate components
