+++ src/ui/components/nav/Nav.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from "react";
-import { UserMenu } from "./components/UserMenu";
+import { UserMenu } from "./components/UserMenu/UserMenu";
 import { CartNavItem } from "./components/CartNavItem";
 import { NavLinks } from "./components/NavLinks";
 import { MobileMenu } from "./components/MobileMenu";
